Improved test coverage + proper preamble support
